refactor(core): model streamable traits for `fuel-core` types directly for simplicity
