+++ crates/fuel-streams-core/src/nats/mod.rs
@@ -17,12 +17,3 @@ pub use nats_client::*;
 pub use nats_conn::*;
 pub use nats_namespace::*;
 pub use streams::{stream::*, subject::*};
-
-pub mod subjects {
-    pub use streams::{
-        stream_blocks::{subjects as blocks, BlockSubjects},
-        stream_transactions::{subjects as transactions, TransactionSubjects},
-    };
-
-    use super::*;
-}
